Precompute isSparseSelection
